Fix issue with server calling wrong domain
