+++ app/api/[address]/register/route.ts
@@ -6,7 +6,7 @@ import {
 } from "@underdog-protocol/types";
 
 import { fetcher } from "@/lib/fetchers";
-import { getBaseUrl } from "@/lib/utils";
+import { getBaseUrl } from "@/lib/server/utils";
 
 export type RegisterResponse = {
     status: string;
