+++ package.json
@@ -50,6 +50,7 @@
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.45.4",
     "react-qr-code": "^2.0.12",
+    "server-only": "^0.0.1",
     "sharp": "^0.32.4",
     "tailwind-merge": "^1.14.0",
     "tailwindcss-animate": "^1.0.6"
